fix: circular dependency error
Resolve #32
